@@ -1,6 +1,6 @@
 module github.com/dinstone/focus-go
 
-go 1.17
+go 1.18
 
 require (
 	github.com/golang/protobuf v1.5.0
