@@ -3,8 +3,6 @@ package test
 import (
 	"errors"
 	"log"
-	"net"
-	"net/rpc"
 	"reflect"
 	"testing"
 
@@ -23,16 +21,15 @@ func init() {
 
 // test client synchronously call
 func client_call(t *testing.T, c compressor.Compressor) {
+	// server
 	server := createServer(c)
 	defer server.Close()
 
 	// client
-	conn, err := net.Dial("tcp", ":8008")
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer conn.Close()
-	client := focus.NewClient(conn, options.WithCompressor(c))
+	x := options.NewClientOptions(":8008")
+	x.SetCompressor(c)
+	x.SetSerializer(serializer.Protobuf)
+	client := focus.NewClient(x)
 	defer client.Close()
 
 	// case
@@ -41,47 +38,52 @@ func client_call(t *testing.T, c compressor.Compressor) {
 		err   error
 	}
 	cases := []struct {
-		client         *focus.Client
-		name           string
-		serviceMenthod string
-		arg            *protobuf.ArithRequest
-		expect         expect
+		client  *focus.Client
+		name    string
+		service string
+		method  string
+		arg     *protobuf.ArithRequest
+		expect  expect
 	}{
 		{
-			client:         client,
-			name:           "test-1",
-			serviceMenthod: "ArithService.Add",
-			arg:            &protobuf.ArithRequest{A: 20, B: 5},
+			client:  client,
+			name:    "test-1",
+			service: "ArithService",
+			method:  "Add",
+			arg:     &protobuf.ArithRequest{A: 20, B: 5},
 			expect: expect{
 				reply: &protobuf.ArithResponse{C: 25},
 				err:   nil,
 			},
 		},
 		{
-			client:         client,
-			name:           "test-2",
-			serviceMenthod: "ArithService.Sub",
-			arg:            &protobuf.ArithRequest{A: 20, B: 5},
+			client:  client,
+			name:    "test-2",
+			service: "ArithService",
+			method:  "Sub",
+			arg:     &protobuf.ArithRequest{A: 20, B: 5},
 			expect: expect{
 				reply: &protobuf.ArithResponse{C: 15},
 				err:   nil,
 			},
 		},
 		{
-			client:         client,
-			name:           "test-3",
-			serviceMenthod: "ArithService.Mul",
-			arg:            &protobuf.ArithRequest{A: 20, B: 5},
+			client:  client,
+			name:    "test-3",
+			service: "ArithService",
+			method:  "Mul",
+			arg:     &protobuf.ArithRequest{A: 20, B: 5},
 			expect: expect{
 				reply: &protobuf.ArithResponse{C: 100},
 				err:   nil,
 			},
 		},
 		{
-			client:         client,
-			name:           "test-4",
-			serviceMenthod: "ArithService.Div",
-			arg:            &protobuf.ArithRequest{A: 20, B: 5},
+			client:  client,
+			name:    "test-4",
+			service: "ArithService",
+			method:  "Div",
+			arg:     &protobuf.ArithRequest{A: 20, B: 5},
 			expect: expect{
 				reply: &protobuf.ArithResponse{C: 4},
 			},
@@ -89,43 +91,39 @@ func client_call(t *testing.T, c compressor.Compressor) {
 		{
 			client,
 			"test-5",
-			"ArithService.Div",
+			"ArithService",
+			"Div",
 			&protobuf.ArithRequest{A: 20, B: 0},
 			expect{
 				&protobuf.ArithResponse{},
-				rpc.ServerError("divided is zero"),
+				errors.New("(301)divided is zero"),
 			},
 		},
 	}
 	for _, c := range cases {
 		t.Run(c.name, func(t *testing.T) {
 			reply := &protobuf.ArithResponse{}
-			err := c.client.Call(c.serviceMenthod, c.arg, reply)
+			err := c.client.Call(c.service, c.method, c.arg, reply)
 			assert.Equal(t, true, reflect.DeepEqual(c.expect.reply.C, reply.C))
 			assert.Equal(t, c.expect.err, err)
 		})
 	}
 }
 
-func createServer(c compressor.Compressor) net.Listener {
-	// server
-	lis, err := net.Listen("tcp", ":8008")
+func createServer(c compressor.Compressor) *focus.Server {
+	x := options.NewServerOptions(":8008")
+	if c != nil {
+		x.SetCompressor(c)
+	}
+	x.SetSerializer(serializer.Protobuf)
+	server := focus.NewServer(x)
+	err := server.Register(new(protobuf.ArithService))
 	if err != nil {
 		log.Fatal(err)
 	}
-	server := focus.NewServer(options.WithCompressor(c))
-	err = server.Register(new(protobuf.ArithService))
-	if err != nil {
-		log.Fatal(err)
-	}
-	go server.Serve(lis)
+	server.Start()
 
-	return lis
-}
-
-// TestClient_Call test client synchronously call
-func TestClient_Call(t *testing.T) {
-	client_call(t, compressor.Raw)
+	return server
 }
 
 // TestClient_AsyncCall test client asynchronously call
@@ -133,13 +131,9 @@ func TestClient_AsyncCall(t *testing.T) {
 	server := createServer(nil)
 	defer server.Close()
 
-	// client
-	conn, err := net.Dial("tcp", ":8008")
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer conn.Close()
-	client := focus.NewClient(conn)
+	x := options.NewClientOptions(":8008")
+	x.SetSerializer(serializer.Protobuf)
+	client := focus.NewClient(x)
 	defer client.Close()
 
 	type expect struct {
@@ -147,44 +141,49 @@ func TestClient_AsyncCall(t *testing.T) {
 		err   error
 	}
 	cases := []struct {
-		client        *focus.Client
-		name          string
-		serviceMethod string
-		arg           *protobuf.ArithRequest
-		expect        expect
+		client  *focus.Client
+		name    string
+		service string
+		method  string
+		arg     *protobuf.ArithRequest
+		expect  expect
 	}{
 		{
-			client:        client,
-			name:          "test-1",
-			serviceMethod: "ArithService.Add",
-			arg:           &protobuf.ArithRequest{A: 20, B: 5},
+			client:  client,
+			name:    "test-1",
+			service: "ArithService",
+			method:  "Add",
+			arg:     &protobuf.ArithRequest{A: 20, B: 5},
 			expect: expect{
 				reply: &protobuf.ArithResponse{C: 25},
 			},
 		},
 		{
-			client:        client,
-			name:          "test-2",
-			serviceMethod: "ArithService.Sub",
-			arg:           &protobuf.ArithRequest{A: 20, B: 5},
+			client:  client,
+			name:    "test-2",
+			service: "ArithService",
+			method:  "Sub",
+			arg:     &protobuf.ArithRequest{A: 20, B: 5},
 			expect: expect{
 				reply: &protobuf.ArithResponse{C: 15},
 			},
 		},
 		{
-			client:        client,
-			name:          "test-3",
-			serviceMethod: "ArithService.Mul",
-			arg:           &protobuf.ArithRequest{A: 20, B: 5},
+			client:  client,
+			name:    "test-3",
+			service: "ArithService",
+			method:  "Mul",
+			arg:     &protobuf.ArithRequest{A: 20, B: 5},
 			expect: expect{
 				reply: &protobuf.ArithResponse{C: 100},
 			},
 		},
 		{
-			client:        client,
-			name:          "test-4",
-			serviceMethod: "ArithService.Div",
-			arg:           &protobuf.ArithRequest{A: 20, B: 5},
+			client:  client,
+			name:    "test-4",
+			service: "ArithService",
+			method:  "Div",
+			arg:     &protobuf.ArithRequest{A: 20, B: 5},
 			expect: expect{
 				reply: &protobuf.ArithResponse{C: 4},
 			},
@@ -192,23 +191,29 @@ func TestClient_AsyncCall(t *testing.T) {
 		{
 			client,
 			"test-5",
-			"ArithService.Div",
+			"ArithService",
+			"Div",
 			&protobuf.ArithRequest{A: 20, B: 0},
 			expect{
 				&protobuf.ArithResponse{},
-				rpc.ServerError("divided is zero"),
+				errors.New("(301)divided is zero"),
 			},
 		},
 	}
 	for _, cs := range cases {
 		t.Run(cs.name, func(t *testing.T) {
 			reply := &protobuf.ArithResponse{}
-			call := cs.client.AsyncCall(cs.serviceMethod, cs.arg, reply)
+			call := cs.client.AsyncCall(cs.service, cs.method, cs.arg, reply)
 			err := <-call
 			assert.Equal(t, true, reflect.DeepEqual(cs.expect.reply.C, reply.C))
 			assert.Equal(t, cs.expect.err, err.Error)
 		})
 	}
+}
+
+// TestClient_Call test client synchronously call
+func TestClient_Call(t *testing.T) {
+	client_call(t, compressor.Raw)
 }
 
 // TestNewClientWithSnappyCompress test snappy comressor
@@ -228,7 +233,7 @@ func TestNewClientWithZlibCompress(t *testing.T) {
 
 // TestServer_Register .
 func TestServer_Register(t *testing.T) {
-	server := focus.NewServer()
+	server := focus.NewServer(options.ServerOptions{})
 	err := server.RegisterName("ArithService", new(protobuf.ArithService))
 	assert.Equal(t, nil, err)
 	err = server.Register(new(protobuf.ArithService))
@@ -238,25 +243,23 @@ func TestServer_Register(t *testing.T) {
 // TestNewClientWithSerializer .
 func TestNewClientWithSerializer(t *testing.T) {
 	// server
-	lis, err := net.Listen("tcp", ":8010")
+	server := focus.NewServer(options.NewServerOptions(":8010"))
+	err := server.Register(new(json.TestService))
 	if err != nil {
 		log.Fatal(err)
 	}
+	server.Start()
 
-	server := focus.NewServer(options.WithSerializer(serializer.Json))
-	err = server.Register(new(json.TestService))
-	if err != nil {
-		log.Fatal(err)
-	}
-	go server.Serve(lis)
+	// var wg sync.WaitGroup
+	// wg.Add(1)
+	// go func() {
+	// 	// do something
+	// 	wg.Done()
+	// }()
+	// wg.Wait()
 
 	// client
-	conn, err := net.Dial("tcp", ":8010")
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer conn.Close()
-	client := focus.NewClient(conn, options.WithSerializer(serializer.Json))
+	client := focus.NewClient(options.NewClientOptions(":8010"))
 	defer client.Close()
 
 	type expect struct {
@@ -264,17 +267,19 @@ func TestNewClientWithSerializer(t *testing.T) {
 		err   error
 	}
 	cases := []struct {
-		client         *focus.Client
-		name           string
-		serviceMenthod string
-		arg            *json.Request
-		expect         expect
+		client  *focus.Client
+		name    string
+		service string
+		method  string
+		arg     *json.Request
+		expect  expect
 	}{
 		{
-			client:         client,
-			name:           "test-1",
-			serviceMenthod: "TestService.Add",
-			arg:            &json.Request{A: 20, B: 5},
+			client:  client,
+			name:    "test-1",
+			service: "TestService",
+			method:  "Add",
+			arg:     &json.Request{A: 20, B: 5},
 			expect: expect{
 				reply: &json.Response{C: 25},
 			},
@@ -283,7 +288,7 @@ func TestNewClientWithSerializer(t *testing.T) {
 	for _, c := range cases {
 		t.Run(c.name, func(t *testing.T) {
 			reply := &json.Response{}
-			err := c.client.Call(c.serviceMenthod, c.arg, reply)
+			err := c.client.Call(c.service, c.method, c.arg, reply)
 			assert.Equal(t, true, reflect.DeepEqual(c.expect.reply.C, reply.C))
 			assert.Equal(t, c.expect.err, err)
 		})
